+++ spatialaudiometrics/load_data.py
@@ -3,12 +3,14 @@ Load data module
 '''
 import os
 import sys
+import wget
 from importlib import resources
 from pysofaconventions import SOFAFile
 import numpy as np
 import pandas as pd
 from spatialaudiometrics import angular_metrics as am
 from spatialaudiometrics import localisation_metrics as lm
+from spatialaudiometrics import hrtf_metrics as hf
 
 class HRTF:
     '''
@@ -28,6 +30,15 @@ class HRTF:
         self.locs       = np.round(sofa.getVariableValue('SourcePosition').data,2) # Round to avoid any bizarre precision errors
         self.hrir       = sofa.getDataIR().data
         self.fs         = sofa.getSamplingRate().data[0]
+        # Get ITD
+        delay           = sofa.getDataDelay().data
+        if np.shape(delay)[0] == 1:
+            print('Estimating itd using the threshold method')
+            itd_s, itd_samps, itd_index = hf.itd_estimator_threshold(self.hrir,self.fs)
+            self.itd_s = itd_s
+        else:
+            itd             = delay[:,0] - delay[:,1]
+            self.itd_s      = itd/self.fs
 
 def load_example_behavioural_data():
     '''
@@ -46,6 +57,24 @@ def load_example_sofa_files():
         hrtf2 = HRTF(sofa)
     return hrtf1,hrtf2
 
+def load_sonicom_sofa(subject:str,hrir_type:str,sample_rate:int,no_itd:bool = False):
+    '''
+    Loads in a SONICOM sofa file give you know the type of hrir you want and the participant number
+    :param subject: the pnumber of the subject (e.g. P0107)
+    :param hrir_type: type of hrir from this list: Raw, Windowed, FreeFieldComp, FreeFieldCompMinPhase
+    :param sample_rate: sample rate of the hrtf you want to load (44,48, or 96)
+    :param no_itd: A boolean that says if you want the ITD in the hrir (no_itd = False) or you want the version with the ITD removed from the hrir and in the metadata instead (no_itd = True)
+    '''
+    if no_itd:
+        hrir_type = hrir_type + '_NoITD'
+    link = 'ftp://transfer.ic.ac.uk:2122/2022_SONICOM-HRTF-DATASET/'+subject+'/HRTF/HRTF/' + str(sample_rate) + 'kHz/'+subject+'_'+hrir_type+'_' + str(sample_rate) + 'kHz.sofa'
+    temp_filename = 'load_sonicom_sofa_temp_file.sofa'
+    if os.path.exists(temp_filename):
+        os.remove(temp_filename) # if exist, remove it directly
+    wget.download(link,temp_filename)
+    hrtf = HRTF(temp_filename)
+    hrtf.sofa_path = link
+    return hrtf
 
 def match_hrtf_locations(hrtf1,hrtf2):
     '''
@@ -71,9 +100,11 @@ def match_hrtf_locations(hrtf1,hrtf2):
         sys.exit('Error: Was not able to match all the locations in hrtf 1 with hrtf 2')
     hrtf2.hrir          = hrtf2.hrir[loc2_idx,:,:]
     hrtf2.locs          = hrtf2.locs[loc2_idx,:]
+    if len(hrtf2.itd_s) > 1:
+        hrtf2.itd_s         = hrtf2.itd_s[loc2_idx]
     return hrtf1, hrtf2
 
-def preprocess_behavioural_data(df:pd.DataFrame):
+def preprocess_behavioural_data(df:pd.DataFrame,cone_size_degrees = 45):
     '''
     Preprocesses the data to make sure we generate the information needed to calculate the metrics
     
@@ -113,7 +144,7 @@ def preprocess_behavioural_data(df:pd.DataFrame):
                                                               df.at[i,'azi_response'],df.at[i,'ele_response'])
         
         # Classify confusions
-        df.at[i,'confusion_classification'] = lm.classify_confusion(df.iloc[i])
+        df.at[i,'confusion_classification'] = lm.classify_confusion(df.iloc[i],cone_size_degrees=cone_size_degrees)
 
     # Calculate weighting
     df['polar_weight']           = lm.polar_error_weight(df)
@@ -131,3 +162,4 @@ def preprocess_behavioural_data(df:pd.DataFrame):
     df['unsigned_polar_error']   = abs(df.signed_polar_error)
 
     return df
+