+++ spatialaudiometrics/visualisation.py
@@ -5,11 +5,14 @@ import seaborn as sns
 import numpy as np
 import matplotlib
 import matplotlib.pyplot as plt
+import matplotlib.colors as colors
+import scipy
+from mpl_toolkits.axes_grid1.inset_locator import inset_axes
+from matplotlib.lines import Line2D
+import matplotlib.animation as animation
 from spatialaudiometrics import hrtf_metrics as hf
 from spatialaudiometrics import angular_metrics as am
-from mpl_toolkits.axes_grid1.inset_locator import inset_axes
-from matplotlib.cm import ScalarMappable
-import matplotlib.animation as animation
+from spatialaudiometrics import signal_processing as sp
 
 
 class FigSize:
@@ -41,6 +44,19 @@ class FontSize:
     axis_labels     = 9
     default         = 10
     fig_title       = 12
+
+class Colours:
+    dict = {  
+        'L':colors.hex2color('#196AA5'),
+        'R':colors.hex2color('#C75E6B')
+        }
+
+class Palettes:
+    '''
+    Custom palettes for use in seaborn plotting
+    '''
+    left_right = sns.color_palette([Colours.dict['L'],Colours.dict['R']])
+    confusion = {"precision":"black","front-back":"red","in-cone":"blue","off-cone":"orange"}
 
 def create_fig(fig_size = (FigSize.fig_width,FigSize.fig_height),grid_spec_rows = 12, grid_spec_cols = 12):
     '''
@@ -151,11 +167,9 @@ def plot_error_bar(axes,df,x_name,y_name,sample_name):
     '''
 
     x       = np.arange(0,len(df[x_name].unique()),1) + 0.2
-    av      = df.groupby(x_name).mean(numeric_only = True).reset_index()
-    std     = df.groupby(x_name).std(numeric_only = True).reset_index()/(len(df[sample_name].unique())**0.5)
-    y       = av[y_name]
-    yerr    = std[y_name]
-    axes.errorbar(x, y, yerr, fmt='o', color = [0.3,0.3,0.3], linewidth = 3, markersize = 7)
+    y       = df.groupby(x_name).mean(numeric_only = True).reset_index()[y_name]
+    std     = df.groupby(x_name).std(numeric_only = True).reset_index()[y_name]/(len(df[sample_name].unique())**0.5)
+    axes.errorbar(x, y, std, fmt='o', color = [0.3,0.3,0.3], linewidth = 3, markersize = 7)
 
 def annotate_p_vals(axes,pval,x_coord):
     '''
@@ -272,6 +286,7 @@ def plot_itd_overview(hrtf):
     
     fig,gs  = create_fig(fig_size=(16,6))
     axes    = fig.add_subplot(gs[0:12,0:4], projection = 'polar')
+    axes.set_theta_direction(-1)
     axes.plot(np.deg2rad(hrtf.locs[idx,0]),np.abs(itd_us[idx]))
     axes.set_theta_zero_location("N")
     axes.set_rticks([200,400,600,800])
@@ -298,6 +313,7 @@ def plot_ild_overview(hrtf):
 
     fig,gs  = create_fig(fig_size=(16,6))
     axes    = fig.add_subplot(gs[0:12,0:4], projection = 'polar')
+    axes.set_theta_direction(-1)
     axes.plot(np.deg2rad(hrtf.locs[idx,0]),np.abs(ild[idx]))
     axes.set_theta_zero_location("N")
     #axes.set_rticks([200,400,600,800])
@@ -329,18 +345,18 @@ def add_colourbar_on_side(hue_min,hue_max,colourmap,axes,axes_label):
     cbar    = axes.figure.colorbar(sm, ax = axes)
     cbar.set_label(axes_label, rotation = 90)
 
-def plot_source_locations(locs):
+def plot_source_locations(locs,elevation_palette = "tab10", point_size = 25):
     '''
     Plots the source locations on a 3d plot.
     
     :param locations: numpy array where each row is a location, col 1 = azimuth, col 2 = elevation, col 3 = distance. Can easily just use hrtf.locs
     '''
     # Plot the locations used
-    fig,gs = create_fig(fig_size=(10,10))
+    fig,gs = create_fig(fig_size=(5,5))
     x,y,z = am.polar2cartesian(locs[:,0], locs[:,1], locs[:,2])
     dist = max(abs(z))
     axes = fig.add_subplot(gs[0:12,0:12],projection='3d')
-    axes.scatter(x,y,z,s = 100)
+    axes.scatter(x,y,z,s = point_size,c = z, cmap = elevation_palette)
     axes.set_xlabel('X')
     axes.set_ylabel('Y')
     axes.set_title('Source locations')
@@ -350,11 +366,11 @@ def plot_source_locations(locs):
     axes.set_aspect('equal')
     # Annotate
     x,y,z = am.polar2cartesian(0, 0, locs[0,2])
-    axes.text(x,y,z,  'Front', size=14, zorder=1, color = 'k')
+    axes.text(x,y,z,  'Front', size=14, color = 'k',)
     x,y,z = am.polar2cartesian(90, 0, locs[0,2])
-    axes.text(x,y,z,  'Left', size=14, zorder=1, color = 'k')
+    axes.text(x,y,z,  'Left', size=14, color = 'k')
     x,y,z = am.polar2cartesian(270, 0, locs[0,2])
-    axes.text(x,y,z,  'Right', size=14, zorder=1, color = 'k')
+    axes.text(x,y,z,  'Right', size=14, color = 'k')
     finish_axes(axes)
     show()
     return fig, axes
@@ -375,3 +391,237 @@ def create_source_location_gif(fig,axes,save_filename,dpi = 120):
     rot_animation = animation.FuncAnimation(fig, rotate, frames=np.arange(0, 362, 2), interval=100)
     rot_animation.save(save_filename, dpi=dpi, writer='pillow')
     print("Saved animation at: " + save_filename)
+
+
+def plot_hrir_both_ears(hrtf,az,el,axes):
+    '''
+    Plots the hrir at a given position for left and right ears
+    '''
+    # Get location
+    idx = np.where((hrtf.locs[:,0] == az) & (hrtf.locs[:,1] == el))[0][0]
+
+    hrir_l = hrtf.hrir[idx,0,:]
+    hrir_r = hrtf.hrir[idx,1,:]
+
+    ts = (np.arange(0,len(hrir_l))/hrtf.fs)*1000
+    axes.plot(ts,hrir_l,color = Colours.dict['L'])
+    axes.plot(ts,hrir_r,color = Colours.dict['R'])
+    axes.set_ylabel('Amplitude')
+    axes.set_xlabel('Time (ms)')
+    axes.set_title('Azimuth: ' + str(az) + '°, elevation: ' + str(el) + '°.')
+    finish_axes(axes)
+
+def plot_hrtf_both_ears(hrtf,az,el,axes,log_scale = True):
+    '''
+    lots the transfer function at a given position for left and right ears
+    '''
+    idx = np.where((hrtf.locs[:,0] == az) & (hrtf.locs[:,1] == el))[0][0]
+
+    hrtfs, freqs, hrtfs_phase = hf.hrir2hrtf(hrtf.hrir,hrtf.fs,db_flag = 1)
+
+    hrtf_l = hrtfs[idx,0,:]
+    hrtf_r = hrtfs[idx,1,:]
+
+    axes.plot(freqs/1000,hrtf_l,color = Colours.dict['L'])
+    axes.plot(freqs/1000,hrtf_r,color = Colours.dict['R'])
+    axes.set_ylabel('Magnitude (dB)')
+    axes.set_xlabel('Frequency (kHz)')
+    axes.set_title('Azimuth: ' + str(az) + '°, elevation: ' + str(el) + '°.')
+    if log_scale:
+        axes.set_xscale('log')
+        axes.set_xticks([0.125,0.25,0.5,1,2,4,8,16,22])
+        axes.set_xticklabels([0.125,0.25,0.5,1,2,4,8,16,22])
+
+    finish_axes(axes)
+
+def plot_spectrogram(sig,fs,axes,vmin_max = False, freq_max = 20000, db_flag = 0):
+    '''
+    Plots the spectrogram of a signal using wavelet decomposition
+    :param sig: 1D signal you want to run the spectrogram on
+    :param fs: sample rate
+    :param axes: axes you want to plot the spectrogram on
+    :vmin_max: a tuple with the min and max values you want to display
+    :param freq_max: the maximum frequency you want to calculate for
+    :param db_flag: if you want the spectrogram to display in magnitude (0) or dB (1)
+    :return freqs: frequencies
+    :return mag: magnitude or dB of signal at each frequency
+    :return phase: phase of the signal at each frequency
+    '''
+    #print('Running wavelet decomposition...')
+    #mag,phase,freqs = sp.wavelet_decomposition(sig,fs,freq_steps = 10, freq_max = freq_max)
+
+    # Just using the scipy one for ow
+    freqs, ts, mag = scipy.signal.spectrogram(sig, fs)
+
+    if db_flag == 1:
+        mag = sp.mag2db(mag)
+
+    print('Plotting...')
+    #ts = np.arange(0,len(sig)/fs,1/fs)
+    if vmin_max:
+        c = axes.pcolorfast(ts,freqs/1000,mag,cmap = 'mako',vmin = vmin_max[0],vmax = vmin_max[1])
+    else:
+        c = axes.pcolorfast(ts,freqs/1000,mag,cmap = 'mako')
+
+    axes.set_xlabel('Time (s)') 
+    axes.set_ylabel('Frequency (kHz)')
+    #axes.set_yscale('log')
+    cax = inset_axes(axes,width="2.5%",height="100%",bbox_transform=axes.transAxes,bbox_to_anchor=(0.025, 0.05, 1.05, 0.95),loc= 1)
+    plt.colorbar(c,cax = cax)
+
+    return freqs,mag  #phase
+
+
+def plot_spectrum(sig,fs,axes):
+    '''
+    Plots the spectrum
+    '''
+    spec, freqs, phase = sp.calculate_spectrum(sig,fs)
+    axes.plot(freqs/1000,spec,'k')
+    axes.set_xlabel('Frequency (kHz)')
+    axes.set_ylabel('Magnitude (dB)')
+    finish_axes(axes)
+
+def plot_ild_itd_difference(df,diverging = True):
+    '''
+    Creates a plot to show the itd and ild difference
+    '''
+    # Plot ITD difference
+    fig,gs = create_fig()
+    axes = fig.add_subplot(gs[1:6,1:12])
+    if diverging:
+        palette = "vlag"
+        huemax = max(abs(df.itd_diff_us))
+        huemin = huemax
+    else:
+        palette = "viridis"
+        huemin = min(df.itd_diff_us)
+        huemax = max(df.itd_diff_us)
+
+    sns.scatterplot(data=df, x="az", y="el", hue = "itd_diff_us", hue_norm = (-huemin,huemax), ax = axes, palette = palette)
+    axes.set_title('ITD difference')
+    axes.set_ylabel('Elevation (°)')
+    axes.set_xlabel('Azimuth (°); -> counterclockwise')
+    finish_axes(axes)
+    add_colourbar_on_side(-huemin,huemax,palette,axes,'ITD difference (μs)')
+
+    # Plot ILD difference
+    if diverging:
+        huemax = max(abs(df.ild_diff_db))
+        huemin = huemax
+    else:
+        huemin = min(df.ild_diff_db)
+        huemax = max(df.ild_diff_db)
+
+    axes = fig.add_subplot(gs[7:12,1:12])
+    sns.scatterplot(data=df, x="az", y="el", hue = "ild_diff_db", hue_norm = (-huemin,huemax), ax = axes, palette = palette)
+    axes.set_title('ILD difference')
+    axes.set_ylabel('Elevation (°)')
+    axes.set_xlabel('Azimuth (°); -> counterclockwise')
+    finish_axes(axes)
+    add_colourbar_on_side(-huemin,huemax,palette,axes,'ILD difference (dB)')
+    show()
+    return fig
+
+def plot_LSD_left_and_right(df):
+    '''
+    Plot on the same plot of the log spectral distortion on the L and R given a data frame with it already generated
+    '''
+    fig,gs = create_fig()
+    maxLSD = max([max(abs(df.lsd_l)),max(abs(df.lsd_r))])
+    minLSD = min([min(abs(df.lsd_l)),min(abs(df.lsd_r))])
+    axes = fig.add_subplot(gs[1:6,1:12])
+    sns.scatterplot(data=df, x="az", y="el", hue_norm = (minLSD,maxLSD), hue = "lsd_l", ax = axes, palette = "PuRd")
+    axes.set_title('LSD of left ear')
+    axes.set_ylabel('Elevation (°)')
+    axes.set_xlabel('Azimuth (°); -> counterclockwise')
+    finish_axes(axes)
+    add_colourbar_on_side(minLSD,maxLSD,"PuRd",axes,'Log spectral distortion (dB)')
+
+    axes = fig.add_subplot(gs[7:12,1:12])
+    sns.scatterplot(data=df, x="az", y="el", hue_norm = (minLSD,maxLSD), hue = "lsd_r", ax = axes, palette = "PuRd")
+    axes.set_title('LSD of right ear')
+    axes.set_ylabel('Elevation (°)') 
+    axes.set_xlabel('Azimuth (°); -> counterclockwise')
+    finish_axes(axes)
+    add_colourbar_on_side(minLSD,maxLSD,"PuRd",axes,'Log spectral distortion (dB)')
+    show()
+    return fig
+
+def plot_LSD_left_right_frequency(df):
+    '''
+    Plot the LSD as a function of frequency
+    '''
+    fig,gs = create_fig(fig_size = (8,4))
+    df['freqs'] = df['freqs']/1000
+    axes = sns.lineplot(data = df, x = "freqs", y = "lsd", hue = "ear",palette = Palettes.left_right, errorbar = "sd")
+    axes.set_title("Log spectral distortion across location for each frequency (synthetic/measured)")
+    axes.set_ylabel('Log spectral distortion (dB)')
+    axes.set_xlabel('Frequency (kHz)')
+    axes.set_xlim([0,24])
+    axes.set_xscale('log')
+    freqs = [0.25,0.5,1,2,4,8,12,16,20]
+    axes.set_xticks(freqs)
+    axes.set_xticklabels(freqs)
+    axes.set_xlim([0.2,20])
+    finish_axes(axes,legend = 1)
+
+    return fig
+
+def plot_raw_localisation(df,axes,coord):
+    '''
+    Plots a scatterplot with target vs. response
+    :param df: data frame to plot - that has the coord_target and also the confusion classification
+    :param axes: subplot axes to plot in
+    :param coord: what coordinate do you want to plot? (e.g. azi, ele, lat, pol)
+    '''
+    sns.scatterplot(data = df, x = coord + '_target', y = coord + '_response',alpha = 0.5, hue = 'confusion_classification',s = 40 ,ax = axes,palette = Palettes.confusion)
+    axes.set_ylabel('Response')
+    axes.set_xlabel('Target')
+    axes.set_box_aspect(1)
+    finish_axes(axes,grid = 2)
+
+def create_raw_localisation_legend(axes):
+    # Need to do a manual legend
+    legend_elements = [Line2D([0], [0], marker='o', color='w', label='precision', markerfacecolor='black', markersize=10,alpha = 0.5),
+                    Line2D([0], [0], marker='o', color='w', label='front-back', markerfacecolor='red', markersize=10,alpha = 0.5),
+                    Line2D([0], [0], marker='o', color='w', label='in-cone', markerfacecolor='blue', markersize=10,alpha = 0.5),
+                    Line2D([0], [0], marker='o', color='w', label='off-cone', markerfacecolor='orange', markersize=10,alpha = 0.5)]
+    axes.legend(handles = legend_elements)
+
+def plot_confusion_sphere(df,azi_target:float,ele_target:float, point_size = 25):
+    '''
+    Plots the response to a specific target in a sphere and colours by confusion
+    :param df: Dataframe of the preprocessed behavioural data 
+    :param azi_target: Azimuth coordinate of the target to filter by
+    :param ele_taret: Elevation coordinate of the target to filter by
+    :param point_size: Size of points on sphere
+    '''
+    df              = df.loc[((df.azi_target == azi_target) & (df.ele_target == ele_target))]
+    distance        = 1.5
+    df['distance']  = distance # only here for the visualisation
+    fig,gs          = create_fig(fig_size=(5,5))
+
+    x,y,z           = am.polar2cartesian(df.azi_response, df.ele_response, df.distance)
+    dist            = max(abs(z))
+    axes            = fig.add_subplot(gs[0:12,0:12],projection='3d')
+    
+    axes.scatter(x,y,z,s = point_size,c = [Palettes.confusion[i] for i in df.confusion_classification],alpha = 0.5)
+    axes.set_xlabel('X')
+    axes.set_ylabel('Y')
+    axes.set_title('Source locations')
+    axes.set_ylim(-dist,dist)
+    axes.set_xlim(-dist,dist)
+    axes.set_zlim(-dist,dist)
+    axes.set_aspect('equal')
+    # Annotate
+    x,y,z = am.polar2cartesian(0, 0, distance)
+    axes.text(x,y,z,  'Front', size=14, color = 'k',)
+    x,y,z = am.polar2cartesian(90, 0, distance)
+    axes.text(x,y,z,  'Left', size=14, color = 'k')
+    x,y,z = am.polar2cartesian(270, 0, distance)
+    axes.text(x,y,z,  'Right', size=14, color = 'k')
+    finish_axes(axes)
+    create_raw_localisation_legend(axes)
+    show()
+    return fig, axes